Minor reorganization
