+++ firefox-ios/Client/Frontend/Browser/SearchEngines/SearchEnginesManager.swift
@@ -8,6 +8,8 @@ import Shared
 import Storage
 
 struct SearchEngineFlagManager {
+    /// Whether Search Engine Consolidation is enabled.
+    /// If enabled, search engines are fetched from Remote Settings rather than our pre-bundled XML files.
     static var isSECEnabled: Bool {
         // return LegacyFeatureFlagsManager.shared.isFeatureEnabled(.searchEngineConsolidation, checking: .buildOnly)
         // SEC always disabled (for now)
